+++ tests/integration/profiles/valid/es5-polyfilled-promise.server.js
@@ -1,0 +1,9 @@
+function LocalPromise(executor) {
+  executor(function () {});
+}
+
+LocalPromise.resolve = function (value) {
+  return value;
+};
+Promise = LocalPromise;
+const ready = Promise.resolve(1);